# testdata/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Etherniti REST API Proxy definition",
    "version": "0.0.4",
    "title": "Etherniti REST API Proxy",
    "termsOfService": "https://www.etherniti.org/docs/terms-of-service/",
    "contact": {
      "email": "core@etherniti.org"
    },
    "license": {
      "name": "Etherniti License",
      "url": "https://www.etherniti.org/docs/license/"
    },
    "x-apisguru-categories": [
      "iot",
      "ethereum",
      "blockchain"
    ],
    "x-logo": {
      "url": "/favicon.png",
      "backgroundColor": "#FFFFFF"
    },
    "x-preferred": true,
    "x-providerName": "etherniti.org"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/v1",
  "consumes": [
    "application/json;charset=utf-8"
  ],
  "produces": [
    "application/json;charset=utf-8"
  ],
...
  "schemes": ["http", "https", "unix"],
  "paths": {
...
    "/public/profile": {
      "post": {
        "tags": ["profile"],
        "summary": "Create a connection profile using prefered provider",
        "description": "Create a connection profile using prefered provider. Choose between, public node, private node, infura, etc",
        "operationId": "profile",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "deprecated": false,
        "parameters": [{
          "in": "body",
          "required": true,
          "name": "profile",
          "schema": {
            "example": {
              "endpoint": "http://172.168.0.2",
              "address": "0xE720c341bd664A8F7ebC3f75CCE93BD6ba6d7654"
            }
          }
        }],
        "responses": {
          "200": {
            "description": "profile creation response",
            "schema": {
              "example": {
                "id": 0,
                "code": 200,
                "msg": "profile token successfully created",
                "result": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbmRwb2ludCI6Imh0dHA6Ly8xNzIuMTY4LjAuMiIsImFkZHJlc3MiOiIweEU3MjBjMzQxYmQ2NjRBOEY3ZWJDM2Y3NUNDRTkzQkQ2YmE2ZDc2NTQiLCJrZXkiOiIiLCJ2ZXJzaW9uIjoxLCJ2YWxpZGl0eSI6ZmFsc2UsImV4cCI6MTU1NDQ5MzM0MywianRpIjoiZTMzZjhkNTUtM2YzNS00OWQwLTg5ZTgtOWNiMjMzYThiYmQzIiwiaWF0IjoxNTU0MTMzMzQzLCJpc3MiOiJldGhlcm5pdGkub3JnIiwibmJmIjoxNTU0MTMzMzQzfQ.__b0kA8JnwLv0ceI03LRDzt8m3pemoIf_8_0GX76MJg"
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "connectionProfileToken": {
      "type": "apiKey",
      "description": "Connection profile configuration token for Etherniti REST API Proxy",
      "name": "X-Etherniti-Profile",
      "in": "header"
    }
  },
...
}